fix: 쿠폰 생성 수정
