+++ src/main/java/com/nhnacademy/booklay/booklayfront/controller/admin/coupon/CouponAdminFrontController.java
@@ -3,7 +3,6 @@ package com.nhnacademy.booklay.booklayfront.controller.admin.coupon;
 import static com.nhnacademy.booklay.booklayfront.dto.coupon.ControllerStrings.ADMIN_COUPON_REST_PREFIX;
 import static com.nhnacademy.booklay.booklayfront.dto.coupon.ControllerStrings.ATTRIBUTE_NAME_MEMBER_NO;
 import static com.nhnacademy.booklay.booklayfront.dto.coupon.ControllerStrings.DOMAIN_PREFIX_COUPON;
-import static com.nhnacademy.booklay.booklayfront.dto.coupon.ControllerStrings.DOMAIN_PREFIX_SHOP;
 import static com.nhnacademy.booklay.booklayfront.dto.coupon.ControllerStrings.PAGE_NUM;
 import static com.nhnacademy.booklay.booklayfront.dto.coupon.ControllerStrings.RETURN_ADMIN_PAGE;
 import static com.nhnacademy.booklay.booklayfront.dto.coupon.ControllerStrings.TARGET_VIEW;
@@ -11,25 +10,17 @@ import static com.nhnacademy.booklay.booklayfront.utils.ControllerUtil.buildStri
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nhnacademy.booklay.booklayfront.dto.coupon.CouponAddRequest;
-import com.nhnacademy.booklay.booklayfront.dto.storage.response.ObjectFileResponse;
 import com.nhnacademy.booklay.booklayfront.service.RestService;
+import com.nhnacademy.booklay.booklayfront.service.coupon.CouponAdminService;
 import com.nhnacademy.booklay.booklayfront.service.restapimodelsetting.CouponRestApiModelSettingService;
 import java.io.IOException;
-import java.net.URI;
 import java.util.Map;
 import java.util.Objects;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.io.ByteArrayResource;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.client.MultipartBodyBuilder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -51,6 +42,7 @@ public class CouponAdminFrontController {
     private final String gatewayIp;
     private final RestTemplate restTemplate;
 
+    private final CouponAdminService couponAdminService;
     private final CouponRestApiModelSettingService couponRestApiModelSettingService;
 
     private final String COUPON_RESOURCE_BASE = "admin/coupon/";
@@ -84,31 +76,11 @@ public class CouponAdminFrontController {
 
         Map<String, Object> map = objectMapper.convertValue(couponAddRequest, Map.class);
 
-        // 이미지 저장
         if (!Objects.isNull(image) && !image.isEmpty()) {
-            URI storageUrl = URI.create(gatewayIp + DOMAIN_PREFIX_SHOP + "/storage");
-            ByteArrayResource contentsAsResource = new ByteArrayResource(image.getBytes()) {
-                @Override
-                public String getFilename() {
-                    return image.getOriginalFilename();
-                }
-            };
-            MultipartBodyBuilder resource = new MultipartBodyBuilder();
-            resource.part("file",
-                    contentsAsResource,
-                    MediaType.valueOf(Objects.requireNonNull(image.getContentType())));
-
-            MultiValueMap<String, HttpEntity<?>> multipartBody = resource.build();
-
-            HttpHeaders headers = new HttpHeaders();
-            headers.setContentType(MediaType.MULTIPART_FORM_DATA);
-            HttpEntity<MultiValueMap<String, HttpEntity<?>>> httpEntity = new HttpEntity<>(multipartBody,
-                    headers);
-
-            ResponseEntity<ObjectFileResponse> responseEntity = restTemplate.postForEntity(storageUrl, httpEntity,
-                    ObjectFileResponse.class);
-            map.put("fileId", responseEntity.getBody().getId());
+            Long fileId = couponAdminService.saveCouponImage(image);
+            map.put("fileId", fileId);
         }
+        // 이미지 저장
         String url = buildString(gatewayIp, DOMAIN_PREFIX_COUPON, ADMIN_COUPON_REST_PREFIX);
 
         restService.post(url, map, String.class);
