+++ src/main/java/com/nhnacademy/booklay/booklayfront/service/coupon/CouponAdminService.java
@@ -1,0 +1,8 @@
+package com.nhnacademy.booklay.booklayfront.service.coupon;
+
+import java.io.IOException;
+import org.springframework.web.multipart.MultipartFile;
+
+public interface CouponAdminService {
+    Long saveCouponImage(MultipartFile image) throws IOException;
+}
