--- src/main/resources/templates/admin/coupon/empty.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-
-
-    <div id="header-carousel" class="carousel slide" data-ride="carousel">
-
-    </div>
-</html>
